requirements.txt
mechanize
browser-cookie3
